refactor: LottoNumbers > Lotto 객체로 통합
