+++ src/main/java/lotto/domain/WinningLotto.java
@@ -22,8 +22,4 @@ public class WinningLotto extends Lotto {
     public Rank getRank(Lotto lotto) {
         return Rank.valueOf(getCorrectCount(lotto), lotto.hasBonusBall(bonusBall));
     }
-
-    private int getCorrectCount(Lotto lotto) {
-        return this.numbers.getCorrectCount(lotto.numbers);
-    }
 }
